Animation Instance, Animation Data, Animation Data Asset
In this the following steps were taken to implement data-driven animations:

A custom anim instance class was created.
An animation data structure was defined with references to the blank space and idle animations.
A data asset was created to hold the animation data structure.
The data asset was added as a field to the character data asset.
Getters were added to the animal instance class to retrieve the animation data fields from the character data asset.
The parent of the existing animation blueprint was changed to the newly created one.
The blend space and idle animation functions were exposed as pins.
A default animation reference was added to the instance to prevent null returns.
The reference was set and recompiled.
The editor typos were fixed.
The end result allows for easy setup of character animations by simply switching the animation data asset, which can be useful for having different locomotion animations for different character types or driven by equipped weapons.