Footsteps Component, Step Anim Notify
In this , I implemented a Footsteps component in Unreal Engine. We created the component class and added it to our character as the default subject. We also added an enum to specify whether it's the left or right footstep.

To test if the footstep happened and on which surface, we performed a line trace and used the physical material to play the corresponding footstep sound. We also added a console variable for debugging purposes, which shows the name of the material if debugging is turned on.

To trigger the handle footstep function in the component, we created a notify that gets the character from the MASH component and uses a gather to access the component.

Finally, we added footstep notifies to our animations and added footstep sound references to our materials. We also used the safe function to handle non-identified materials.

Overall, these steps allowed us to implement a functional and robust Footsteps component in our Unreal Engine project.