Jump Ability
In this section, I implemented the ability to jump in the Unreal Engine. We started by creating a base class for the ability, which allowed us to wrap the existing jump functionality and manage its relationship to other abilities. We then created the ability blueprint, added the condition to run, and defined the exact activation of the ability by calling the jump function on the character.

We also created an effect that was applied to the character during the jump. This effect was useful for assigning gameplay cues or modifying attributes during the jump. We added tags to our ability, such as the stack tag and state tag, which allowed us to distinguish it from other abilities and track the jump in state.

To trigger the ability, we sent a gameplay event and added a trigger to it. We then assigned the exact tags to the character and tested the ability by jumping. We noticed that the state tag was missing, which we fixed by setting the effect duration to infinite.

Overall, we successfully implemented the ability to jump in the Unreal Engine, which will be useful for further gameplay mechanics.