Base Ability
Created two effect arrays, one for applying effects on start and another for tracking effects during the ability's lifetime.
Implemented activate and end functions to handle the effects.
Created an effect context to be used by the effects and iterated over the array to check for valid effects, create an effects pack, and apply the effects to self.
Saved effect handles to the array for tracking purposes and removed effects by handles on ability end.
Implemented a function to get the game-based character from the actor info, which is callable from blueprints and is blueprint pure.
Set the effect on the jump ability to the new array and cleared the old variable.
Removed all defect definitions and cleared the saved array of effect handles on ability end.