Wall Run Ability (Ability Tasks)
To implement wall run, we will need all run animations. If you have good animations, you can use those. Otherwise, you can create a simple one and mirror it.

Let's create our C++ classes. The first class we will create is our wall run ability class. The second class will be the ability task class, which will handle the wall running itself.

We will start implementation from our ability test class this time. First of all, let's add some delegates. Delegates is the way how we can communicate with the ability to ask how we can get some information from it.

The first delegate will tell us that the wall side of the wall running is determined, and it will return one parameter, a Boolean variable, which will tell us whether it's left or not. The next delegate is the one you might want to use for any task. It will tell us that the task has finished its job, and we have to do something after it. In our case, it will trigger when we can no longer continue the wall run.

Let's create a folder for it. We will create a few variables to cache the variables we're going to pass in that static function. So our character movement component and our character. One instance is going to be created, we're going to set these variables inside the static function on the instance of a test that we're going to create. And of course, the array of objects that we cannot trace against.

Now let's create the function that will create an instance of our task. Every task has to have this function, and it is a static function, so it can be used from code like this or it will be visible for us in blueprints. We will just call it "Create Wall Run Task". The parameters we create, we're going to pass from the ability when are we going to be creating this task? We're going to need the character movement component and trace object types array that will determine which kind of objects we can trace during the wall run. So on this objects of this type, we can wall run, and we will detect them using our traces.

The activate function is where we're going to try to perform the initial determination of the wall. We're going to run on, so let's get our acceleration. Let's not forget the super call. In this function, we're going to try to perform initial determination of the wall. Let's create a few variables to cache the variables we're going to pass in that static function.

The find Runnable wall function will execute for us mostly on tick and will try to find a wall where we can run. If it will return False, it means that right now we no longer have any wall, and we have to finish the wall run. We also need a function that will tell us from which side the wall is. Let's just call it "is wall on the left". We're going to pass the hit result to it.

The task function is going to be our update. The super call, then let's create a hit result for our wall hit because we're going to have to try to determine to find it on every tick. Then the code similar to the one we used on the activation. We probably just can copy it here. In case we fail, we just end the task and broadcast the delegate.

The ability class is going to be a very special ability for us because this time we won't start the logic on the ability activate function, but we're going to start doing our stuff starting from it's been granted. Our goal is going to be to create our wall run task, set its parameters, and start it. We will subscribe to our capsule component hit event, and on hit, we're going to try to activate ourselves.

To summarize, we need to create our C++ classes, add delegates, create variables for delegates, create a function to create an instance of our task, create the activate function, implement the find Runnable wall function, create a hit result for our wall hit, implement the task function, and finally, create the ability class.