Item Static Data (UClass CDO approach), Action Game Statics
In this I created static data class in Unreal Engine's Gameplay Ability System.
The approach recommended is to create a class for each item that will be inherited from a UObject and marked as a blueprint type and label.
The class will define basic parameters of the item that will never change, such as the item's name.
A Helper Blueprint Function Library is created to allow access to the constant variables of the item static data class from blueprints.
A function is created in the Helper Library to return the default object of the specified item static data class.
The function is marked as constant and blueprint pure to make it more user-friendly.
The default object is always available and contains basic parameters of the item that can be accessed from blueprints.
A demonstration is provided showing how to use the Helper Library and item static data class to get an item's name in a blueprint.