Item Entry, Item List, Item Instance
Created A new folder was created for the inventory in the source code.
An inventory record list class was created and added to the inventory folder.
An inventory item instance class was created with a static data class property and an equipped property, both of which are replicated.
A virtual callback function was added to the inventory item instance class, which will be used for trigger events on equipped items.
An inventory list item class was created with a pointer to an inventory item instance.
An inventory list class was created, which derives from the VastArraySerializer and uses a T-array of inventory list items.
Add and remove item functions were added to the inventory list class, which creates new instances of inventory items and removes existing ones based on their type.
The inventory list class was marked for replication and given a replicated fast serialized array property to store the inventory list items.
Basic implementations were made for the functions and properties added to the inventory item instance, inventory list item, and inventory list classes.