Inventory Component
To implement the inventory component, an actor component called "InventoryComponent" can be created. The component can be added to the character class, assuming that the inventory exists on the character and that all items are lost when the character dies.

The inventory component can be set up by overriding the InitializeComponent() function to initialize the starting data, such as the starting items, for debugging purposes. The component can also be marked as replicated, and the "ReplicateSubobjects()" function can be overridden to replicate the inventory item list.

To replicate the inventory list, the function can iterate over the list and replicate each item instance using the "ChannelReplicateSubobject()" function, along with the replicated flags. The inventory list can be added to the "LifetimeReplicatedProps()" function in the character class.

To print the items in the inventory, a debug variable can be added to the inventory component, and the OnTick() function can print the inventory to the log if the variable is set. The item static data can be obtained from the item instance, and a function can be created to return the item static data from the item instance.

Default items can be added to the inventory component by storing them as an array of item static data subclasses, which can be added to the inventory list in the InitializeComponent() function using the "AddItem()" function.